requirements.txt
lxml==4.9.3
python-decouple==3.8
pyspark==3.3.0
apache-airflow-providers-apache-spark
requests==2.31.0
pandas